Refactor TR-55 utilities
Move helper functions for TR-55 into their own module. It
helps reduce the noise in tasks.py, ensuring that a higher
percentage of methods there are actual Celery tasks.
